+++ project.sh
@@ -18,5 +18,4 @@ rm project/analysis.yaml
 #  --parallel --functions-only --incremental --min-lines 5
 #.venv/bin/redup scan . --format toon --output ./project --parallel --incremental --functions-only
 .venv/bin/vallm batch . --recursive --output toon > ./project/validation.toon
-#.venv/bin/vallm batch . --recursive --no-imports --no-complexity
 #vallm batch . --recursive --no-imports --no-complexity --output toon > ./project/validation.toon